+++ src/api/v1/controller/features/businessHour.ts
@@ -7,7 +7,18 @@ import config from '../../config/development';
 
 export const addBusinessHour = async (req: Request, res: Response) => {
     try {
-        const userId: string = res.locals.jwt.userId;
+        // const userId: string = res.locals.jwt.userId;
+        let userId:any; 
+        const type = req.query.type; //type = business, user, null
+        if (type && type === config.businessType) {
+            userId = req.query.userId;
+        } else {
+            userId = res.locals.jwt.userId;
+        }
+        if (!userId || userId === "" || userId === undefined) {
+            return apiResponse.errorMessage(res, 401, "User Id is required!");
+        }
+
         const createdAt = utility.dateWithFormat();
 
         const deleteQuery = `DELETE FROM business_hours WHERE user_id = ${userId}`;
@@ -43,7 +54,17 @@ export const addBusinessHour = async (req: Request, res: Response) => {
 
 export const businessHourList = async (req: Request, res: Response) => {
     try {
-        const userId = res.locals.jwt.userId;
+        // const userId = res.locals.jwt.userId;
+        let userId:any; 
+        const type = req.query.type; //type = business, user, null
+        if (type && type === config.businessType) {
+            userId = req.query.userId;
+        } else {
+            userId = res.locals.jwt.userId;
+        }
+        if (!userId || userId === "" || userId === undefined) {
+            return apiResponse.errorMessage(res, 401, "User Id is required!");
+        }
 
         const sql = `SELECT * FROM business_hours WHERE user_id = ${userId}`;
         const [rows]: any = await pool.query(sql);
